[remove] route 통합 및 삭제
