+++ src/app/api/study-board/icon/route.ts
